+++ app.js
@@ -1,43 +1,12 @@
 const [command, key, value] = process.argv.slice(2);
-const { readPasswords, writePasswords } = require("./lib/passwords");
-
-function get() {
-  try {
-    const passwords = readPasswords();
-    console.log(key, passwords[key]);
-  } catch (error) {
-    console.error(error);
-  }
-}
-
-function set() {
-  console.log("Called SET", key, value);
-  try {
-    const passwords = readPasswords();
-    passwords[key] = value;
-    writePasswords(passwords);
-  } catch (error) {
-    console.error(error);
-  }
-}
-
-function unset() {
-  console.log("Called UNSET", key);
-  try {
-    const passwords = readPasswords();
-    delete passwords[key];
-    writePasswords(passwords);
-  } catch (error) {
-    console.error(error);
-  }
-}
+const { get, set, unset } = require("./lib/commands");
 
 if (command === "get") {
-  get();
+  get(key);
 } else if (command === "set") {
-  set();
+  set(key, value);
 } else if (command === "unset") {
-  unset();
+  unset(key);
 } else {
   console.log("Unkown command");
 }
